Initial commit
This adds the basic link-header based pagination for Django REST
Framework along with basic tests to ensure that it is still
working.

Because pagination serializers cannot modify the headers, this
requires that a mixin is included with the view that will override
the generic pagination.  This relies on a pagination serializer
which defines the "next" and "previous" fields in the response,
similar to what the default pagination serializer does.  This also
includes support for pagination serializers which define "first"
and "last" as well.

The tests ensure that the headers are added for the first, middle,
and last pages.  This should cover the basic tests needed to check
that the previous and next links are added in the correct
situations.

The tests can be run using

    python manage.py test rest_link_pagination

from within a Django project or running

    python rest_link_pagination/tests/run.py

from within the DRF Link Pagination project.